Only report errors that occurred in the input file
This distinguishes errors that occurred in the input file, which might
indicate a real problem, and errors that occurred in the generated
constants for macros. These are basically guaranteed to occur since it
just guesses the type and some macros contain something other than
constants.

Also cleared up the indentation from the last patch.