Pointerize to avoid pass-by-value
The pass-by-value convention is more complicated and may not be
supported by all users of the C library. So certain types need to be
changed to pointers, and then changed dereferenced in the
implementation. This patch also fixes the test case, and fixes a warning
by making typedefs of a type refer to the same struct name.